samples/python/agents/vertex_image_gen/pyproject.toml
[project]
name = "a2a-sample-agent-vertex-image-gen"
version = "0.1.0"
description = "Vertex AI Image Generation Agent with A2A Protocol integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-samples",
    "click>=8.1.8",
    "google-cloud-aiplatform>=1.36.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-samples = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"